Use `serde` instead of `rustc-serialize`
Closes https://github.com/PistonDevelopers/piston/issues/1173

- Moved serialize tests to “tests” folder
