+++ src/input/Cargo.toml
@@ -15,6 +15,10 @@ name = "input"
 path = "src/lib.rs"
 
 [dependencies]
-rustc-serialize = "0.3.16"
+serde_derive = "1.0"
+serde = "1.0"
 bitflags = "0.9.1"
 piston-viewport = "0.3.0"
+
+[dev-dependencies]
+serde_json = "1.0"
